Run 8 details:
This is the first run that incorporates a bunch of recent changes to the rate rules for the ROO reactions. Also,
in this run, I have set the maximum radical count to 1 (with the fix to allow for O2), so hopefully the mechanism
generation will be faster/less memory intensive. The R. + O2 rates in RMG were overestimated by several orders of
magnitude and more reasonable estimates are now in place, so more of the R+O2 flux should lead to ROO. formation.
The primary oxidation model is not being used as a seed for this run.